requirements.txt
PyQt5
QScintilla
pyserial
